Fix mixing VALUES and RowBinary inserts
According to the offsets file, each file persisted is used either
with VALUES or with RowBinary. We can't mix both within the same
file.

Thus, we can't use the same key for the map independently of the
RowBinary param because then we'd attempt to insert with the wrong
format. Surprisingly, the fact that RowBinary format needs a different
file was already considered, but not for the hash map.